- [x] 1. Terms and conditions вылезают каждый раз при открытии даже после подтверждения - [x] 2. При переключении между Sell и Buy не меняется название колонки с ценой - [x] 3. Проценты отклонения должны окрашиваться в зависимости от стороны отклонения - [x] 4. В поле Date сигнала нужно отображать дату суток закрытия, на которую отображается значение, а не время добавления из админки - [x] 8. На детальном окне сигнала едет верстка (см. скриншот) - [x] 9. Кнопка Add to watchlist не работает - [x] 10. На вкладке Sectors не работает возможность просмотра сигналов по секторам при нажатии

+++ app/src/main/java/com/example/cityfalcon/FiltersForSignlsBottomSheet.java
@@ -75,8 +75,8 @@ public class FiltersForSignlsBottomSheet extends BottomSheetDialogFragment {
     }
 
     //получить id инструмента
-    private Float instrumentId;
-    private Float GetInstrumentId (String filterOrInstrumentTitle){
+    private Integer instrumentId;
+    private Integer GetInstrumentId (String filterOrInstrumentTitle){
         RetrofitCreate.getRetrofit().GetInstrumentId(registrationResponse.getAccept(),
                 registrationResponse.getAuthorization(),
                 filterOrInstrumentTitle).enqueue(new Callback<InstrumentArticle>() {
